First pass at using mixins in Pitop and AlexRobot
