+++ pitop/robotics/drive_controller.py
@@ -1,10 +1,14 @@
-from math import floor, pi
+from math import (
+    floor,
+    pi,
+    radians,
+)
+from time import sleep
 
 from pitop.pma import (
     EncoderMotor,
     ForwardDirection,
 )
-
 from pitop.system.port_manager import PortManager
 from simple_pid import PID
 
@@ -73,7 +77,16 @@ class DriveController:
         self._left_motor.set_target_rpm(target_rpm=rpm_left)
         self._right_motor.set_target_rpm(target_rpm=rpm_right)
 
-    def forward(self, speed_factor, hold):
+    def forward(self, speed_factor, hold=False):
+        """
+        Move the robot forward.
+
+        :param float speed_factor:
+            Factor relative to the maximum motor speed, used to set the velocity, in the range -1.0 to 1.0.
+            Using negative values will cause the robot to move backwards.
+        :param bool hold:
+            Setting this parameter to true will cause subsequent movements to use the speed set as the base speed.
+        """
         linear_speed_x = self._max_motor_speed * speed_factor
         if hold:
             self._linear_speed_x_hold = linear_speed_x
@@ -81,20 +94,59 @@ class DriveController:
             self._linear_speed_x_hold = 0
         self.__robot_move(linear_speed_x, 0)
 
-    def backward(self, speed_factor, hold):
+    def backward(self, speed_factor, hold=False):
+        """
+        Move the robot backward.
+
+        :param float speed_factor:
+            Factor relative to the maximum motor speed, used to set the velocity, in the range -1.0 to 1.0.
+            Using negative values will cause the robot to move forwards.
+        :param bool hold:
+            Setting this parameter to true will cause subsequent movements to use the speed set as the base speed.
+        """
         self.forward(-speed_factor, hold)
 
-    def left(self, speed_factor, turn_radius):
+    def left(self, speed_factor, turn_radius=0):
+        """
+        Make the robot move to the left, using a circular trajectory.
+
+        :param float speed_factor:
+            Factor relative to the maximum motor speed, used to set the velocity, in the range -1.0 to 1.0.
+            Using negative values will cause the robot to turn right.
+        :param float turn_radius:
+            Radius used by the robot to perform the movement. Using `turn_radius=0` will cause the robot to rotate in place.
+        """
+
         self.__robot_move(self._linear_speed_x_hold, self._max_robot_angular_speed * speed_factor, turn_radius)
 
-    def right(self, speed_factor, turn_radius):
+    def right(self, speed_factor, turn_radius=0):
+        """
+        Make the robot move to the right, using a circular trajectory.
+
+        :param float speed_factor:
+            Factor relative to the maximum motor speed, used to set the velocity, in the range -1.0 to 1.0.
+            Using negative values will cause the robot to turn left.
+        :param float turn_radius:
+            Radius used by the robot to perform the movement. Using `turn_radius=0` will cause the robot to rotate in place.
+        """
+
         self.left(-speed_factor, -turn_radius)
 
     def target_lock_drive_angle(self, angle):
         angular_speed = self.__target_lock_pid_controller(angle)
         self.__robot_move(self._linear_speed_x_hold, angular_speed)
 
-    def rotate(self, angle, angular_speed):
+    def rotate(self, angle, time_to_take):
+        """
+        Rotate the robot in place by a given angle and stop.
+
+        :param float angle: Angle of the turn.
+        :param float time_to_take: Expected duration of the rotation, in seconds.
+        """
+        assert time_to_take > 0.0
+        angle_radians = radians(angle)
+        angular_speed = angle_radians / time_to_take
+
         angular_speed = angular_speed * angle / abs(angle)
         rpm_left, rpm_right = self.__calculate_motor_rpms(0, angular_speed, turn_radius=0)
         rotations = abs(angle) * pi * self._wheel_separation / (360 * self._wheel_circumference)
@@ -102,12 +154,24 @@ class DriveController:
                                         total_rotations=rotations*rpm_left/abs(rpm_left))
         self._right_motor.set_target_rpm(target_rpm=rpm_right,
                                          total_rotations=rotations*rpm_right/abs(rpm_right))
+        sleep(time_to_take)
 
     def stop(self):
+        """
+        Completely stops the robot.
+        """
         self._linear_speed_x_hold = 0
         self.__robot_move(0, 0)
 
     def stop_rotation(self):
+        """
+        Stops any angular movement performed by the robot.
+
+        In the case where linear and rotational movements are being performed at
+        the same time (e.g.: during a left turn with a turn radius different to 0),
+        calling this method will cause the robot to continue the linear movement,
+        so it will continue to move forward.
+        """
         self.__robot_move(self._linear_speed_x_hold, 0)
 
     def _speed_to_rpm(self, speed):
